Basic Static Hazard Implemented
Added a Hazard script and a static hazard (the floating cube over on the
right of the Physics Test stage) that makes use of it. The hazard should
cause damage and knockback, but since they haven't been implemented yet
it only sends a debug message.
I also moved the call to FindHurtboxes in Character's Awake function to
after movementCollider and triggerCollider are assigned, and added
Character as a Hurtbox layer in Config. These were both necessary to
make the hazard work properly.